+++ src/Querify/IFetchRequest.cs
@@ -1,0 +1,8 @@
+﻿using System.Linq;
+
+namespace Querify
+{
+    public interface IFetchRequest<TQueried, TFetch> : IOrderedQueryable<TQueried>
+    {
+    }
+}